Set a default timeout for run_script
It would appear that popen is not blocking the process return without a pipe.

The communicate() should be doing this, but we might be hitting a bug in popen?

Therefore, set a default timeout as a workaround. The better solution would be to
figure out why we aren't blocking. I have mentioned that as a FIXME.