Check for file path type before calling os.remove()
When running the test suite with Python 3, we need to check for the
type of a file path before calling os.remove(). In PY3, os.path.exists()
returns true as there are still open file descriptors at this point,
but tmp_file.name is an integer instead of a string, causing os.remove to
stacktrace.

This change checks for the the tmp_file.name type and closes the file handle.