Handle the decode function better when reading in the spooled tmp file
Refs #72

When reading in the spooled tmp file, we must decode the binary data
immediately. However, we must handle this differently in Py2 and 3.
For Py3, we will try to use salt's system encoding dunder if possible,
and then fall back to utf-8 for now. In Py2, we're not reading for
everything to be unicode yet in Salt, so we'll keep these as string
objects until we can make a full switch.

Thanks to @s0undt3ch for the pointers here to not decode straight to
ascii inadvertently.