Fix typo in run_run_plus, add fun name and jid to return
This fixes a typo as well as adds the jid to the return (to be used in
some tests I am about to write where I need to check the job cache after
a runner is executed).

I am making similar changes in the run_run_plus func in
tests/integration/__init__.py on the salt repo.