Fix incorrect version comparison
In #73 we introduced a check against the python version.

However, this is written incorrectly and will return true for any
version of Python. It's supposed to return true only for versions
that are 3.x or above but instead it returns true for any version of python
that's 2.x or above.

ex for Python 2.7.12:

>>> sys.version_info > (2,)
True
>>> sys.version_info > (3,)
False

This breaks Unicode handling for many cases where a test uses a shell
and gets unicode in the response.